Move browser-driven GC-cursor test out of the browserless build job
test_disconnect_releases_gc_cursor_promptly launches chromium, but the
`build` CI job runs `pytest --ignore=tests/e2e` WITHOUT installing a
browser -- so it failed there ("Executable doesn't exist"), and the e2e
job (which runs only tests/e2e/) never ran it at all. Moved to
tests/e2e/test_broadcast_gc_disconnect.py: excluded from build, run by
the e2e job with chromium. The no-browser replay-marker test stays in
tests/test_message_buffer.py.

This is why the pytest workflow was red on the branch; local full-suite
runs passed only because they have chromium installed.
